Add model data validation

+++ model/post.js
@@ -1,6 +1,7 @@
 const User = require('../model/user');
 const Product = require('../model/product');
 const db = require('../db/mysql');
+const validator = require('../public/js/backendValidator');
 
 class Post {
     //parametr id jest na końcu, bo jest opcjonalny
@@ -12,11 +13,13 @@ class Post {
     }
 
     //dodawanie obiektu do bazy
-    static add(text,userId,productId) {
-        return db.execute(
-            'insert into Post (Text, IdProduct, IdUser) values (?, ?, ?)',
-            [text, productId, userId]
-          );
+    static add(text, userId, productId) {
+        if (validator.isValid(text) && validator.isValid(userId) && validator.isValid(productId)) {
+            return db.execute(
+                'insert into Post (Text, IdProduct, IdUser) values (?, ?, ?)',
+                [text, productId, userId]
+            );
+        }
     }
 
     //pobranie listy obiektów
@@ -28,37 +31,47 @@ class Post {
     }
 
     static listUserPosts(idUser) {
-        return db.execute('SELECT Post.IdPost, Post.Text, Users.Picture as UserPicture, Users.IdUser as UserId, Users.FirstName as UserFirstName, Users.LastName as UserLastName, Product.Name as ProductName, Product.Volume as ProductVolume, Product.Capacity as ProductCapacity, Product.Price as ProductPrice FROM Post INNER JOIN Users ON Users.IdUser=Post.IdUser INNER JOIN Product ON Product.IdProduct=Post.IdProduct where Users.IdUser = ?;',
-        [idUser]);
+        if (validator.isValid(idUser)) {
+            return db.execute('SELECT Post.IdPost, Post.Text, Users.Picture as UserPicture, Users.IdUser as UserId, Users.FirstName as UserFirstName, Users.LastName as UserLastName, Product.Name as ProductName, Product.Volume as ProductVolume, Product.Capacity as ProductCapacity, Product.Price as ProductPrice FROM Post INNER JOIN Users ON Users.IdUser=Post.IdUser INNER JOIN Product ON Product.IdProduct=Post.IdProduct where Users.IdUser = ?;',
+                [idUser]);
+        }
     }
 
     static listProductPosts(idProduct) {
-        return db.execute('select Post.IdPost, Post.Text, Users.FirstName as UserFirstName, Users.LastName as UserLastname, Users.Picture UserPicture, Users.DateOfBirth as UserDateOfBirth from Post LEFT JOIN Users ON Users.IdUser=Post.IdUser where IdProduct = ?;',
-        [idProduct]);
+        if (validator.isValid(idProduct)) {
+            return db.execute('select Post.IdPost, Post.Text, Users.FirstName as UserFirstName, Users.LastName as UserLastname, Users.Picture UserPicture, Users.DateOfBirth as UserDateOfBirth from Post LEFT JOIN Users ON Users.IdUser=Post.IdUser where IdProduct = ?;',
+                [idProduct]);
+        }
     }
 
     //edycja obiektu
-    static edit(id,text) {
-        return db.execute(
-            'update Post set Text = ? where IdPost = ?',
-            [text, id]
-          );
+    static edit(id, text) {
+        if (validator.isValid(id) && validator.isValid(text)) {
+            return db.execute(
+                'update Post set Text = ? where IdPost = ?',
+                [text, id]
+            );
+        }
     }
 
     //usuwanie obiektu po id
-    static delete(id,userId) {
-        return db.execute(
-            'delete from Post where IdPost = ? and IdUser = ?',
-            [id,userId]
-          );
+    static delete(id, userId) {
+        if (validator.isValid(id) && validator.isValid(userId)) {
+            return db.execute(
+                'delete from Post where IdPost = ? and IdUser = ?',
+                [id, userId]
+            );
+        }
     }
 
     //pobieranie obiektu do widoku szczegółów
     //może być potrzebne pobranie dodatkowych danych
     //np. przez złączenia JOIN w relacyjnej bazie danych
     static details(id) {
-        return db.execute('SELECT Post.IdPost, Post.Text, Users.IdUser as UserId, Users.Picture as UserPicture,  Users.FirstName as UserFirstName, Users.LastName as UserLastName, Product.IdProduct as ProductId, Product.Name as ProductName, Product.Volume as ProductVolume, Product.Capacity as ProductCapacity, Product.Price as ProductPrice FROM Post INNER JOIN Users ON Users.IdUser=Post.IdUser  INNER JOIN Product ON Product.IdProduct=Post.IdProduct where Post.IdPost = ?;',
-      [id]);
+        if (validator.isValid(id)) {
+            return db.execute('SELECT Post.IdPost, Post.Text, Users.IdUser as UserId, Users.Picture as UserPicture,  Users.FirstName as UserFirstName, Users.LastName as UserLastName, Product.IdProduct as ProductId, Product.Name as ProductName, Product.Volume as ProductVolume, Product.Capacity as ProductCapacity, Product.Price as ProductPrice FROM Post INNER JOIN Users ON Users.IdUser=Post.IdUser  INNER JOIN Product ON Product.IdProduct=Post.IdProduct where Post.IdPost = ?;',
+                [id]);
+        }
     }
 }
