+++ model/producer.js
@@ -1,6 +1,7 @@
 const User = require('../model/user');
 const Product = require('../model/product');
 const db = require('../db/mysql');
+const validator = require('../public/js/backendValidator');
 
 class Producer {
     //parametr id jest na końcu, bo jest opcjonalny
@@ -14,10 +15,16 @@ class Producer {
 
     //dodawanie obiektu do bazy
     static add(producer) {
-        return db.execute(
-            'insert into Producer (Name, Country, Founded_Date, Owner) values (?, ?, ?, ?)',
-            [producer.name, producer.country, producer.dateOfStart, producer.owner]
-          );
+        if (validator.isValid(producer.name) &&
+            validator.isValid(producer.country) &&
+            validator.isDateValid(producer.dateOfStart) &&
+            validator.isValid(producer.owner)
+        ) {
+            return db.execute(
+                'insert into Producer (Name, Country, Founded_Date, Owner) values (?, ?, ?, ?)',
+                [producer.name, producer.country, producer.dateOfStart, producer.owner]
+            );
+        }
     }
 
     //pobranie listy obiektów
@@ -30,25 +37,38 @@ class Producer {
 
     //edycja obiektu
     static edit(producer) {
-        return db.execute(
-            'update Producer set Name = ?, Country = ?, Founded_Date = ?, Owner = ? where IdProducer = ?',
-            [producer.name, producer.country, producer.dateOfStart, producer.owner, producer.id]
-          );
+        if (validator.isValid(producer.name) &&
+            validator.isValid(producer.country) &&
+            validator.isDateValid(producer.dateOfStart) &&
+            validator.isValid(producer.owner) &&
+            validator.isValidId(producer.id)
+        ) {
+            return db.execute(
+                'update Producer set Name = ?, Country = ?, Founded_Date = ?, Owner = ? where IdProducer = ?',
+                [producer.name, producer.country, producer.dateOfStart, producer.owner, producer.id]
+            );
+        }
     }
+
     //usuwanie obiektu po id
     static delete(id) {
-        return db.execute(
-            'delete from Producer where IdProducer = ?;',
-            [id]
-          );
+        if (validator.isValidId(id)) {
+            return db.execute(
+                'delete from Producer where IdProducer = ?;',
+                [id]
+            );
+        }
     }
+
     //pobieranie obiektu do widoku szczegółów
     //może być potrzebne pobranie dodatkowych danych
     //np. przez złączenia JOIN w relacyjnej bazie danych
     static details(id) {
-        return db.execute('select * from Producer where IdProducer = ?',
-      [id]
-      );
+        if (validator.isValidId(id)) {
+            return db.execute('select * from Producer where IdProducer = ?',
+                [id]
+            );
+        }
     }
 }
